requirements.txt
openai==0.27.8
pyjwt
python-dotenv
psycopg2-binary
flask==2.3.2
waitress==2.1.2
flask-cors==4.0.0
datetime
pytz
requests